Use source location for ansible roles when in a venv
It's somewhat common to use python packages like this inside a
venv created by a non-root user[1].  Doing so currently fails as the
packaged data files can't be installed into /usr/share. We can instead
use an editable install[2], but later this will cause problems as the
ansible roles aren't in the expected location.  We also could override
the ANSIBLE_ROLES_PATH envvar as is done in tcib-dev-env.rc.  This works
but lacks discoverability.

Instead we can detect that the code is running inside a venv and add the
source dir to the ANSIBLE_ROLES_PATH

[1]
$ git clone ....
$ cd tcib && python -m venv .venv
$ .venv/bin/pip install .

[2]
$ git clone ....
$ cd tcib && python -m venv .venv
$ .venv/bin/pip install -e .