Add container image update subcommand
The existing build and hotfix subcommands do not handle the case where
we want to update any/all packages installed in the container from an
existing repo.  It's possible that we could mirror a repo to the a
build machine and then use `hotfix` but this is fragile as we may
install new unwanted RPMS.  Using what essentially boils down to
`dnf update` with a restricted set of repos is more robust and flexible.

This change adds a new subcommand
 `openstack tcib container image update` that exposes the existing
yum_update tasks via the command line.