Allow source container-images from the build host, not an rpm
It is currently not practical to do local tcib development and build
images with those changes, because the container-images directory is
sourced from an rpm install inside the container, instead of the
container-images from the build host.

With this change, clearing the `tcib_package` variable will result in
--config-path and TCIB_CONFIG_PATH being honoured for the non-template
files in container-images/kolla.

The README.md has been reworked to describe how to do local development.