remove unused image and pmem deps
nova upstream supports persitent memory but
intel discontinued the product line and downstream
support has been removed from our product.
As a result the daxio and ndctl binaries are nolonger used.

This change removes them and the unused nova-compute-ironic image.
The normal nova-compute image will be used for ironic, libvirt and
the fake driver so the nova-compute-ironic image is not required.

Closes: OSPRH-136