Ceilometer: Remove python3-pyngus
This package was required to use amqp1 backend which is being
deprecated[1].

Usage of amqp1 was required by current architecture of STF. However
the architecture will be updated so that we no longer use qpid dispatch
router(aka QDR). So this requirement will be no longer valid in
operator deployment.

[1] https://review.opendev.org/c/openstack/oslo.messaging/+/861070