Introduce kolla_copy_cacerts
This adds a script to run at container start via kolla_start to
update the container environment for trusted CAs (TLS-E).

Also adds the cloud-admin user to the kolla group to be allowed
to run kolla* commands as root. This is required to get the
openstackclient CA trust updated for tls endpoints.

Jira: OSP-26299
Jira: OSP-26849