Fix dependency handling logic
The note[1] indicates that dependencies will be built but this
only appears to work for the first ancestor of an image.
Fix this by walking up the directory tree for a given image to
find all ancestors to include in the build.

Also record the level of each image and process in a breath first order,
removing the hardcoded ordering for the 'base' and 'os' images.

[1] https://github.com/openstack-k8s-operators/tcib/blob/41a73ab686826c4fc2095cbe3f76b0dc2d61147f/tcib/client/container_image.py#L537