Add tempest parameters for the tempest container
This patch introduces tempest parameters for the tempest container.
Currently, it is not possible to specify with what parameters should
be tempest executed.

This patch introduces a set of environment variables prefixed with
TEMPEST_ that can be used to tell the container which arguments
should be used when executing tempest.