Revert "Revert "Merge pull request #96 from lpiwowar/feature/tempest-parameter""
This patch reverts the revert and contains the fix for an error which
was probably causing the failure of tempest in CI.

The issue probably was the fact that TEMPEST_PATH was set at the
beginning of run_tempest.sh to /var/lib/tempest which made the script
to ignore files in external_files directory.

This patch makes sure that when USE_EXTERNAL_FILES variable is set that
TEMPEST_PATH contains the correct path to external files
(/var/lib/tmepest/external_files).

This reverts commit 54c7d984ed751ddc560c868f1e7769865948d8d6.