remove unused deps from nova-compute container
Over the year the nova compute contiainer image in tripleo
gained a number of depencies that are not used by nova but were used
by tripleo. The also enabled funcitonality that were not supported
in our downstream product.

This change removes openssh as the server is not required in this container,
libguestfs was previously included but file injection has never been supported
downstream and is deprecated and disabled by default upstream.
parted is not used by nova in production code nor is jq so both are
removed.

jq was added due to https://bugzilla.redhat.com/show_bug.cgi?id=2090752
as part of https://bugzilla.redhat.com/show_bug.cgi?id=2109644 that has now been
resolved in rhel and we should not need this workaround anymore so the run command
is removed

comments are added to note why some packages are required to prevent
regressions.

Related: OSPRH-192