[tobiko] install tox version >= 4.13
With this patch, the tobiko image installs a later tox version,
>=4.13, with pip.
That tox version does not require the python library py, which has
vulnerabilities:
https://github.com/advisories/GHSA-w596-4wvx-j9j6

This tox version requires a python3-packaging version later than the one
from the rpm installed on the base/os image. Due to that, the tobiko
image is moved from base/os/tobiko to base/tobiko, so that
python3-packaging is not installed via via rpm.

Depends-On: https://review.opendev.org/c/x/tobiko/+/910589

OSPRH-5225