Remove pysnmp from explicit list of packages in ironic-conductor
During Caracal release, ironic, proliantutils and scciclient have moved
to pysnmp-lextudio fork to replace the unmaintained pysnmp.

In order to make this container file compatible with both antelope and
master, we can symply remove pysnmp from the explicit list of packages
and simply let proliantutils and scciclient to pull the right package
for each release.