Add option to add extra image
This patch adds the option to specify an extra image for tests that
require advanced images. This change is specific for Whitebox
Neutron Tempest plugin as it is not certain, if general solution is
needed. It would be also ineffective to achieve generalisation due
to different methods of notation. So for now, the solution is done
for Whitebox Neutron Tempest plugin only.