[Tobiko] Download tobiko repo, oc and ubuntu image during image creation
The tobiko repo, the openshift client (oc) and the ubuntu image
were downloaded when the tobiko container started running, before this
patch.
After this patch, those operations will be performed during the image
creation, so when the container starts running tobiko, oc and ubuntu
image will already be available.

OSPRH-6347