Make TEMPEST_NEUTRON_IMAGE more generic
The TEMPEST_NEUTRON_IMAGE_* variables were tailored specifically to the
neutron-tempest-plugin. Those variables did not allow to specify things
like flavor, ID or name of the image that should be uploaded to
OpenStack.

This patch introduces a new set of variables called
TEMPEST_EXTRA_IMAGES_* that can be used to specify a list of images
that should be uploaded to OpenStack. These variables accepted
a comma-separated list of values.