[Tempest] Remove cirros image from logs
Up until now we were collecting cirros image that is downloaded during
the discover-tempest-config phase into the logs folder. The cirros image
is not needed artifact and in some edge cases it might happen that the
cirros file is too big for it to be stored in the external PV.

This patch removes the cirros image before the logs are collected.