Check expected failures based on new parameter
By adding new parameter Expected Failures List in the test
operator, we will be able to decide which tests should not have
an impact on the job's pass/fail. This patch introduces a way to
check if the failing tests are expected to fail and in that case
the failing result will be non-voting.