[Test-Operator] Move entrypoint scripts
We want to be able to run test pods spawned by the test-operator with
readOnlyRootFileSystem. This requires mounting of ephemeral storage to
/var/lib/X [1]. In this repository all write actions can be performed
(e.g., installation of additional plugins for tempest).

Mounting of the ephemeral storage causes removal of the run_X.sh [1]
script and other data that are stored in the /var/lib/X directory during
the container build (e.g., tobiko source code directory).  Therefore, we
need to move it to a different location. This patch moves:

-  all the entrypoint scripts for test operator related images
   under /usr/local/bin.

-  all source code downloaded during the build of a container to
   /usr/local/src

-  all images downloaded during the build of a container to
   /usr/local/share

[1] X = [tempest|tobiko|ansibletest|horizontest]

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/2579