Add httpd fix to watcher api containerfile
When trying to run a Watcher API container built with the existing
container file, the httpd server fails with the error
"SSLCertificateFile: file '/etc/pki/tls/certs/localhost.crt'
does not exist or is empty". We already have an script to fix it [1],
but it's not used for the watcher api image, this change adds it.

[1] https://github.com/openstack-k8s-operators/tcib/blob/main/container-images/kolla/base/httpd_setup.sh