Fix expectedFailuresList incorrect overwrite
During a job run, we have noticed that the job status was reported
as success despite an error occurring. The root cause was that
stestr faling list was empty due to an error arising before any
tests were executed. However, expectedFailuresList mistakenly
interpreted this as if there were no failing tests. This patch
aims to fix this edge case.