Always download tempest extra image
Tempest requires a few of the images configured to bot server/volume.
It expects some of the images (image_ref, image_ref_alt) to be present
in glance and tests use glance image id. But in some tests it requires
images to be present in the configured locations. for example:

1. conf.scenario.img_file: https://github.com/openstack/tempest/blob/33c1959cdbedf33a93cce>

In this case, Tempest scenario manager will require the image file to
be present in the configured location and it upload the same to glance.
If it does not find the image in that location (even it is present in glance)
 then it fails:
- https://github.com/openstack/tempest/blob/33c1959cdbedf33a93cce470659f54d2a3f4a160/tempe>

2. New test to boot from iso image: conf.whitebox.http_iso_image
- https://github.com/openstack/whitebox-tempest-plugin/blob/19f39aba511ddf6ecddd0e7dd5ae24>

In this case, test need iso image to be present in the configured location
and if it does not exist then test fail (even image is present in glance)
- https://github.com/openstack/whitebox-tempest-plugin/blob/19f39aba511ddf6ecddd0e7dd5ae24>

These above images are configured in TEMPEST_EXTRA_IMAGES_NAME. To solve the
cases, irrespective of tempest extra image is present in glance, we still
need images to be present in the configured location.

Testing:

The 2nd case is tested in:
- https://gitlab.cee.redhat.com/ci-framework/ci-framework-testproject/-/merge_requests/1464
Test passing fine
- https://sf.apps.int.gpc.ocp-hub.prod.psi.redhat.com/zuul/t/components-integration/build/>

Signed-off-by: Ghanshyam Maan <gmaan@redhat.com>