Add function to re-run failed tests
Life is life and the reality is often disappointing.
When you run over a 1000 tempest tests, you can hit
some single ones that failed due to random reason.
This commit implements a killer feature that would
re-run those tests that failed. This gives a chance
to test the fate and not mark the whole build
as a failure if the problem is a pure-random thing
that never occurs again.