Add neutron-fwaas package to be installed in Neutron-server image
This package provides Firewall-As-A-Service service plugin for Neutron.
Project neutron-fwaas is in the Neutron stadium in upstream and is
revived since couple of cycles. Now we are going to also include it in
the RHOSO releases.
This patch adds installation of the python3-neutron-fwaas package only
on the RHEL images (downstream). The reason for that is that in u/s
there is no python3-neutron-fwaas package from master branch for CentOS
9 stream as on CentOS 9 stream there is  python 3.9 but neutron-fwaas
from master branch requires at least python 3.10 since [1].

[1] https://github.com/openstack/neutron-fwaas/commit/ed98cd8a2c87ccac513acfbf448f4795a18aac08

Closes: #OSPRH-15213

Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>