Consume master packages from main branch
The agreed intent of the TCIB main branch is to consume RDO master
packages and produce containers in a master namespace.

ci-framework variables all default to antelope so they need to be
overridden in the TCIB main job config to use the correct packages.

The job names have -master appended to them to reflect what they are
doing.

Likely a similar change needs to be made to the TCIB antelope branch to
be explicit about what packages to consume. Relying on ci-framework
defaults will become an issue when we need to switch to a new upstream
release.

This change results in tcib-openstack-meta-content-provider-master
passing confirming that the master images are created. However
subsequent jobs are made non-voting because this is the first time
ci-framework is running on master content. Fixing this will be a larger
unrelated effort.

Jira: OSPCIX-1078